requirements.txt
# ── Core ──
fastapi==0.135.1
uvicorn[standard]==0.34.0
python-multipart==0.0.22
python-dotenv==1.1.0

# ── Auth ──
python-jose[cryptography]==3.5.0
bcrypt==4.3.0
passlib[bcrypt]==1.7.4

# ── AI ──
anthropic==0.49.0

# ── EEG Analysis ──
mne==1.9.0
autoreject==0.4.3
scipy==1.15.2
numpy==2.2.3

# ── Data / Viz / Reports ──
pandas==2.2.3
plotly==6.0.1
reportlab==4.4.0
Pillow==11.1.0
nibabel==5.3.2
matplotlib==3.10.1

# ── DB ──
sqlalchemy==2.0.38

# ── Utilities ──
httpx==0.28.1
rich==13.9.4
